Working shopping cart
This is a basic working shopping cart with no jQuery and little CSS it
is just the cart in a working state nothing more.

Had to add some jQuery to get the items number and the button is also
not hikashop it will be used to show the full cart.